# Optimized Anvil parsing