--- a/.gitignore
+++ b/.gitignore
@@ -98,3 +98,5 @@ features.toml
 docs/.vitepress/dist
 docs/.vitepress/cache
 node_modules
+
+.idea/