--- a/pumpkin-world/Cargo.toml
+++ b/pumpkin-world/Cargo.toml
@@ -18,7 +18,9 @@ futures = "0.3"
 
 # Compression
 flate2 = "1.0"
-lz4 = "1.11.1"
+lz4 = "1.28.0"
+yazi = "0.2.0"
+zune-inflate = "0.2.54"
 
 serde.workspace = true
 serde_json = "1.0"
@@ -35,3 +37,15 @@ rand = "0.8.5"
 
 num-traits = "0.2"
 num-derive = "0.4"
+memmap2 = "0.9.5"
+tracing = "0.1.40"
+fastanvil = "0.31.0"
+byteorder = "1.5.0"
+
+
+[dev-dependencies]
+criterion = "0.5.1"
+
+[[bench]]
+name = "world"
+harness = false