--- a/pumpkin-world/src/chunk/mod.rs
+++ b/pumpkin-world/src/chunk/mod.rs
@@ -1,10 +1,10 @@
+use fastnbt::LongArray;
+use pumpkin_core::math::vector2::Vector2;
+use serde::{Deserialize, Serialize};
 use std::cmp::max;
 use std::collections::HashMap;
 use std::ops::Index;
 use std::path::PathBuf;
-use fastnbt::LongArray;
-use pumpkin_core::math::vector2::Vector2;
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
 use crate::{